+++ src/client/java/fr/madu59/obe/client/mixin/blockentity/BlockEntityMixin.java
@@ -24,7 +24,7 @@ public class BlockEntityMixin {
             
             boolean hasBook = newBlockState != null && newBlockState.getValueOrElse(LecternBlock.HAS_BOOK, true);
 
-            ext.shouldSkipBeRendering(!hasBook && SettingsManager.OPTIMISED_LECTERNS.getValue() &&  SettingsManager.MOD_TOGGLE.getValue());
+            ext.shouldSkipRendering(!hasBook && SettingsManager.OPTIMISED_LECTERNS.getValue() &&  SettingsManager.MOD_TOGGLE.getValue());
         }
     }
 }