+++ src/client/java/fr/madu59/obe/client/renderer/misc/RenderModeManager.java
@@ -1,15 +1,19 @@
-package fr.madu59.obe.client.renderer.blockentity.misc;
+package fr.madu59.obe.client.renderer.misc;
 
-import fr.madu59.obe.client.chunk.ChunkTaskHolder;
 import fr.madu59.obe.client.config.SettingsManager;
+import fr.madu59.obe.client.chunk.ChunkTaskHolder;
 import fr.madu59.obe.client.config.Option;
 import fr.madu59.obe.client.registry.Registry;
 import fr.madu59.obe.client.renderer.blockentity.ext.BlockEntityExt;
 import fr.madu59.obe.client.renderer.blockentity.ext.BlockEntityRenderStateExt;
+import fr.madu59.obe.client.renderer.entity.ext.EntityExt;
+import fr.madu59.obe.client.renderer.entity.ext.EntityRenderStateExt;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.blockentity.state.BlockEntityRenderState;
+import net.minecraft.client.renderer.entity.state.EntityRenderState;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.SectionPos;
+import net.minecraft.world.entity.Entity;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 
@@ -27,6 +31,18 @@ public class RenderModeManager {
         }
     }
 
+    public static boolean shouldRenderEntity(EntityRenderState state){
+        return shouldRenderEntity(((EntityRenderStateExt)state).entity());
+    }
+
+    public static <T extends Entity> boolean shouldRenderEntity(T entity){
+        return shouldRenderEntity((EntityExt)entity);
+    }
+
+    public static <T extends Entity> boolean shouldRenderEntity(EntityExt ext){
+        return ext.renderMode() == RenderMode.ENTITY;
+    }
+
     public static boolean shouldRenderEntity(BlockEntityRenderState state){
         return shouldRenderEntity(((BlockEntityRenderStateExt)state).blockEntity());
     }
@@ -40,7 +56,7 @@ public class RenderModeManager {
     }
 
     public static <T extends BlockEntity> boolean shouldRenderEntity(boolean setting, BlockEntityExt ext, BlockEntity be){
-        return ext == null || !be.hasLevel() || !ext.isSupportedBlockEntity() || setting || !SettingsManager.MOD_TOGGLE.getValue();
+        return ext == null || !be.hasLevel() || !ext.isSupported() || setting || !SettingsManager.MOD_TOGGLE.getValue();
     }
 
     public static <T extends BlockEntity> boolean shouldRenderEntity(T be){
@@ -48,11 +64,11 @@ public class RenderModeManager {
     }
 
     public static <T extends BlockEntity> boolean shouldRenderEntity(BlockEntityExt ext, T be){
-        return ext == null || !be.hasLevel() || ext.forceEntity() || !ext.isSupportedBlockEntity() || ext.renderMode() == RenderMode.ENTITY || ext.renderModeDelayed() == RenderMode.ENTITY || ext.renderBoth();
+        return ext == null || !be.hasLevel() || ext.forceEntity() || !ext.isSupported() || ext.renderMode() == RenderMode.ENTITY || ext.renderModeDelayed() == RenderMode.ENTITY || ext.renderBoth();
     }
 
     public static <T extends BlockEntity> boolean shouldRenderEntityFast(BlockEntityExt ext){
-        return ext.forceEntity() || !ext.isSupportedBlockEntity() || ext.renderMode() == RenderMode.ENTITY || ext.renderModeDelayed() == RenderMode.ENTITY || ext.renderBoth();
+        return ext.forceEntity() || !ext.isSupported() || ext.renderMode() == RenderMode.ENTITY || ext.renderModeDelayed() == RenderMode.ENTITY || ext.renderBoth();
     }
 
     public static <T extends BlockEntity> void setRenderModeDelayed(T be, RenderMode mode, BlockPos pos){
@@ -78,7 +94,7 @@ public class RenderModeManager {
     }
 
     public static void updateBlockEntityOnChunkRemesh(BlockEntityExt ext, SectionPos pos){
-        if(!ext.isSupportedBlockEntity()) return;
+        if(!ext.isSupported()) return;
         else if(!SettingsManager.MOD_TOGGLE.getValue()) ext.isEnabled(false);
         else{
             String group = Registry.getGroup(((BlockEntity)ext).getType());