+++ CHANGELOG.md
@@ -1,4 +1,2 @@
 # What's new:
-- Removed custom skulls from the mod's file, they can be added back with [this](https://modrinth.com/resourcepack/obe-x-hypixel-skyblock) resource pack
-- Added custom skulls textures loading from resource packs
-- Reduced mod's size by ~17% from removing the textures for the Hypixel's Skyblock compat and putting them in a separate resource pack
+- Fixed a rendering bug with bells